Revert "oki la ban hang moi"
This reverts commit 21f32af0dadd793f9b2d7a6a677ad6112e5f06ea.

+++ 1.DAL/DomainClass/HoaDonChiTiet.cs
@@ -25,9 +25,9 @@ namespace _1.DAL.DomainClass
         public int? SoLuong { get; set; }
         public int? TrangThai { get; set; }
         [Column(TypeName = "decimal(18, 0)")]
-        public int? DonGia { get; set; }
+        public decimal? DonGia { get; set; }
         [Column(TypeName = "decimal(18, 0)")]
-        public int? ThanhTien { get; set; }
+        public decimal? ThanhTien { get; set; }
 
         [ForeignKey(nameof(IdChiTietGiay))]
         [InverseProperty(nameof(ChiTietGiay.HoaDonChiTiets))]
